scrapers/fn-scrape-jobs/requirements.txt
python-jobspy==1.1.80
azure-storage-blob==12.24.0
azure-identity==1.19.0
